Split closed loop test from can_test.py into integration_test.py
